refactor(step 1,2 미션 패키지): step 1,2 미션 패키지 변경
- step1: study → study.carrace.step1
- step2: study → study.carrace.step2
