+++ src/test/java/study/carrace/step1/SetCollectionTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.carrace.step1;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
