+++ src/test/java/study/carrace/step1/StringTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.carrace.step1;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
